✨ Combine ms365 and azure providers into a microsft provider. The asset discovery still happens separately, however both transports now use the same provider.
