+++ motor/discovery/azure/subscriptions.go
@@ -48,6 +48,5 @@ func (client *Subscriptions) GetSubscriptions() ([]subscriptions.Subscription, e
 			subs = append(subs, *s)
 		}
 	}
-
 	return subs, nil
 }
