+++ resources/packs/ms365/msgraph.go
@@ -5,45 +5,23 @@ import (
 
 	"github.com/cockroachdb/errors"
 	"github.com/microsoft/kiota-abstractions-go/authentication"
+	a "github.com/microsoft/kiota-authentication-azure-go"
 	"github.com/microsoftgraph/msgraph-beta-sdk-go/devicemanagement/devicecompliancepolicies"
 	"github.com/microsoftgraph/msgraph-beta-sdk-go/models"
 	"github.com/microsoftgraph/msgraph-beta-sdk-go/rolemanagement/directory/roleassignments"
-	"go.mondoo.com/cnquery/motor/providers"
-	"go.mondoo.com/cnquery/motor/providers/microsoft/azure"
-	ms365_provider "go.mondoo.com/cnquery/motor/providers/microsoft/ms365"
+	microsoft "go.mondoo.com/cnquery/motor/providers/microsoft"
 	"go.mondoo.com/cnquery/motor/providers/microsoft/ms365/msgraphclient"
 	"go.mondoo.com/cnquery/motor/providers/microsoft/ms365/msgraphconv"
-	"go.mondoo.com/cnquery/motor/vault"
 	"go.mondoo.com/cnquery/resources"
 	"go.mondoo.com/cnquery/resources/packs/core"
 )
-
-func msGraphProvider(t providers.Instance) (*ms365_provider.Provider, error) {
-	switch p := t.(type) {
-	case *ms365_provider.Provider:
-		return p, nil
-	case *azure.Provider:
-		cred := p.Credential()
-		if cred == nil {
-			return nil, errors.New("msgraph resource requires credentials")
-		}
-		cfg := &providers.Config{
-			Backend:     providers.ProviderType_MS365,
-			Options:     p.Options(),
-			Credentials: []*vault.Credential{cred},
-		}
-		return ms365_provider.New(cfg)
-	default:
-		return nil, errors.New("msgraph resource is not supported on this provider")
-	}
-}
 
 func (m *mqlMsgraphBeta) id() (string, error) {
 	return "msgraph.beta", nil
 }
 
 func (m *mqlMsgraphBeta) GetSettings() ([]interface{}, error) {
-	provider, err := msGraphProvider(m.MotorRuntime.Motor.Provider)
+	provider, err := microsoftProvider(m.MotorRuntime.Motor.Provider)
 	if err != nil {
 		return nil, err
 	}
@@ -65,7 +43,7 @@ func (m *mqlMsgraphBetaOrganization) id() (string, error) {
 }
 
 func (m *mqlMsgraphBeta) GetOrganizations() ([]interface{}, error) {
-	provider, err := msGraphProvider(m.MotorRuntime.Motor.Provider)
+	provider, err := microsoftProvider(m.MotorRuntime.Motor.Provider)
 	if err != nil {
 		return nil, err
 	}
@@ -118,7 +96,7 @@ func (a *mqlMsgraphBetaGroup) GetMembers() ([]interface{}, error) {
 }
 
 func (m *mqlMsgraphBeta) GetGroups() ([]interface{}, error) {
-	provider, err := msGraphProvider(m.MotorRuntime.Motor.Provider)
+	provider, err := microsoftProvider(m.MotorRuntime.Motor.Provider)
 	if err != nil {
 		return nil, err
 	}
@@ -163,7 +141,7 @@ func (m *mqlMsgraphBetaUser) id() (string, error) {
 }
 
 func (m *mqlMsgraphBeta) GetUsers() ([]interface{}, error) {
-	provider, err := msGraphProvider(m.MotorRuntime.Motor.Provider)
+	provider, err := microsoftProvider(m.MotorRuntime.Motor.Provider)
 	if err != nil {
 		return nil, err
 	}
@@ -227,7 +205,7 @@ func (m *mqlMsgraphBetaServiceprincipal) id() (string, error) {
 }
 
 func (m *mqlMsgraphBeta) GetServiceprincipals() ([]interface{}, error) {
-	provider, err := msGraphProvider(m.MotorRuntime.Motor.Provider)
+	provider, err := microsoftProvider(m.MotorRuntime.Motor.Provider)
 	if err != nil {
 		return nil, err
 	}
@@ -268,7 +246,7 @@ func (m *mqlMsgraphBetaDomain) id() (string, error) {
 }
 
 func (m *mqlMsgraphBeta) GetDomains() ([]interface{}, error) {
-	provider, err := msGraphProvider(m.MotorRuntime.Motor.Provider)
+	provider, err := microsoftProvider(m.MotorRuntime.Motor.Provider)
 	if err != nil {
 		return nil, err
 	}
@@ -320,7 +298,7 @@ func (m *mqlMsgraphBetaDomaindnsrecord) id() (string, error) {
 }
 
 func (m *mqlMsgraphBetaDomain) GetServiceConfigurationRecords() ([]interface{}, error) {
-	provider, err := msGraphProvider(m.MotorRuntime.Motor.Provider)
+	provider, err := microsoftProvider(m.MotorRuntime.Motor.Provider)
 	if err != nil {
 		return nil, err
 	}
@@ -376,7 +354,7 @@ func (m *mqlMsgraphBetaApplication) id() (string, error) {
 }
 
 func (m *mqlMsgraphBeta) GetApplications() ([]interface{}, error) {
-	provider, err := msGraphProvider(m.MotorRuntime.Motor.Provider)
+	provider, err := microsoftProvider(m.MotorRuntime.Motor.Provider)
 	if err != nil {
 		return nil, err
 	}
@@ -425,7 +403,7 @@ func (m *mqlMsgraphBetaUser) GetSettings() (interface{}, error) {
 		return nil, err
 	}
 
-	provider, err := msGraphProvider(m.MotorRuntime.Motor.Provider)
+	provider, err := microsoftProvider(m.MotorRuntime.Motor.Provider)
 	if err != nil {
 		return nil, err
 	}
@@ -501,7 +479,7 @@ func msSecureScoreToMql(runtime *resources.Runtime, score models.SecureScoreable
 }
 
 func (m *mqlMsgraphBetaSecurity) GetLatestSecureScores() (interface{}, error) {
-	provider, err := msGraphProvider(m.MotorRuntime.Motor.Provider)
+	provider, err := microsoftProvider(m.MotorRuntime.Motor.Provider)
 	if err != nil {
 		return nil, err
 	}
@@ -539,7 +517,7 @@ func (m *mqlMsgraphBetaSecurity) GetLatestSecureScores() (interface{}, error) {
 
 // see https://docs.microsoft.com/en-us/graph/api/securescore-get?view=graph-rest-1.0&tabs=http
 func (m *mqlMsgraphBetaSecurity) GetSecureScores() ([]interface{}, error) {
-	provider, err := msGraphProvider(m.MotorRuntime.Motor.Provider)
+	provider, err := microsoftProvider(m.MotorRuntime.Motor.Provider)
 	if err != nil {
 		return nil, err
 	}
@@ -582,7 +560,7 @@ func (s *mqlMsgraphBetaPolicies) id() (string, error) {
 }
 
 func (m *mqlMsgraphBetaPolicies) GetAuthorizationPolicy() (interface{}, error) {
-	provider, err := msGraphProvider(m.MotorRuntime.Motor.Provider)
+	provider, err := microsoftProvider(m.MotorRuntime.Motor.Provider)
 	if err != nil {
 		return nil, err
 	}
@@ -611,7 +589,7 @@ func (m *mqlMsgraphBetaPolicies) GetAuthorizationPolicy() (interface{}, error) {
 }
 
 func (m *mqlMsgraphBetaPolicies) GetIdentitySecurityDefaultsEnforcementPolicy() (interface{}, error) {
-	provider, err := msGraphProvider(m.MotorRuntime.Motor.Provider)
+	provider, err := microsoftProvider(m.MotorRuntime.Motor.Provider)
 	if err != nil {
 		return nil, err
 	}
@@ -636,7 +614,7 @@ func (m *mqlMsgraphBetaPolicies) GetIdentitySecurityDefaultsEnforcementPolicy() 
 
 // https://docs.microsoft.com/en-us/graph/api/adminconsentrequestpolicy-get?view=graph-rest-beta
 func (m *mqlMsgraphBetaPolicies) GetAdminConsentRequestPolicy() (interface{}, error) {
-	provider, err := msGraphProvider(m.MotorRuntime.Motor.Provider)
+	provider, err := microsoftProvider(m.MotorRuntime.Motor.Provider)
 	if err != nil {
 		return nil, err
 	}
@@ -661,7 +639,7 @@ func (m *mqlMsgraphBetaPolicies) GetAdminConsentRequestPolicy() (interface{}, er
 // https://docs.microsoft.com/en-us/azure/active-directory/manage-apps/configure-user-consent?tabs=azure-powershell
 // https://docs.microsoft.com/en-us/graph/api/permissiongrantpolicy-list?view=graph-rest-1.0&tabs=http
 func (m *mqlMsgraphBetaPolicies) GetPermissionGrantPolicies() (interface{}, error) {
-	provider, err := msGraphProvider(m.MotorRuntime.Motor.Provider)
+	provider, err := microsoftProvider(m.MotorRuntime.Motor.Provider)
 	if err != nil {
 		return nil, err
 	}
@@ -688,7 +666,7 @@ func (m *mqlMsgraphBetaRolemanagement) id() (string, error) {
 }
 
 func (m *mqlMsgraphBetaRolemanagement) GetRoleDefinitions() (interface{}, error) {
-	provider, err := msGraphProvider(m.MotorRuntime.Motor.Provider)
+	provider, err := microsoftProvider(m.MotorRuntime.Motor.Provider)
 	if err != nil {
 		return nil, err
 	}
@@ -739,7 +717,7 @@ func (m *mqlMsgraphBetaRolemanagementRoledefinition) id() (string, error) {
 }
 
 func (m *mqlMsgraphBetaRolemanagementRoledefinition) GetAssignments() ([]interface{}, error) {
-	provider, err := msGraphProvider(m.MotorRuntime.Motor.Provider)
+	provider, err := microsoftProvider(m.MotorRuntime.Motor.Provider)
 	if err != nil {
 		return nil, err
 	}
@@ -801,7 +779,7 @@ func (m *mqlMsgraphBetaDevicemanagement) id() (string, error) {
 }
 
 func (m *mqlMsgraphBetaDevicemanagement) GetDeviceConfigurations() ([]interface{}, error) {
-	provider, err := msGraphProvider(m.MotorRuntime.Motor.Provider)
+	provider, err := microsoftProvider(m.MotorRuntime.Motor.Provider)
 	if err != nil {
 		return nil, err
 	}
@@ -848,7 +826,7 @@ func (m *mqlMsgraphBetaDevicemanagement) GetDeviceConfigurations() ([]interface{
 }
 
 func (m *mqlMsgraphBetaDevicemanagement) GetDeviceCompliancePolicies() ([]interface{}, error) {
-	provider, err := msGraphProvider(m.MotorRuntime.Motor.Provider)
+	provider, err := microsoftProvider(m.MotorRuntime.Motor.Provider)
 	if err != nil {
 		return nil, err
 	}
@@ -911,13 +889,14 @@ func (m *mqlMsgraphBetaDevicemanagementDevicecompliancepolicy) id() (string, err
 	return m.Id()
 }
 
-func graphBetaClient(t *ms365_provider.Provider) (*msgraphclient.GraphServiceClient, error) {
-	auth, err := t.Auth()
+func graphBetaClient(t *microsoft.Provider) (*msgraphclient.GraphServiceClient, error) {
+	auth, err := t.GetTokenCredential()
 	if err != nil {
 		return nil, err
 	}
+
 	providerFunc := func() (authentication.AuthenticationProvider, error) {
-		return auth, err
+		return a.NewAzureIdentityAuthenticationProviderWithScopes(auth, msgraphclient.DefaultMSGraphScopes)
 	}
 	adapter, err := msgraphclient.NewGraphRequestAdapterWithFn(providerFunc)
 	if err != nil {
